+++ test/boost/schema_registry_test.cc
@@ -11,6 +11,7 @@
 
 #include <seastar/testing/test_case.hh>
 #include <seastar/testing/thread_test_case.hh>
+#include "data_dictionary/user_types_metadata.hh"
 #include "schema_registry.hh"
 #include "schema_builder.hh"
 #include "test/lib/mutation_source_test.hh"
