+++ test/lib/mutation_source_test.hh
@@ -8,7 +8,8 @@
 
 #pragma once
 
-#include "test/lib/sstable_utils.hh"
+#include "readers/flat_mutation_reader_fwd.hh"
+#include "test/lib/simple_schema.hh"
 
 using populate_fn = std::function<mutation_source(schema_ptr s, const std::vector<mutation>&)>;
 using populate_fn_ex = std::function<mutation_source(schema_ptr s, const std::vector<mutation>&, gc_clock::time_point)>;
