+++ test/lib/simple_schema.hh
@@ -18,10 +18,38 @@
 #include "mutation_fragment.hh"
 #include "mutation.hh"
 #include "schema_builder.hh"
-#include "sstable_utils.hh"
 #include "reader_permit.hh"
 #include "types/map.hh"
 #include "atomic_cell_or_collection.hh"
+
+struct local_shard_only_tag { };
+using local_shard_only = bool_class<local_shard_only_tag>;
+
+//
+// Make set of keys sorted by token for current or remote shard.
+//
+std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size, std::optional<shard_id> shard);
+std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size = 1, local_shard_only lso = local_shard_only::yes);
+
+inline std::vector<sstring> make_keys_for_shard(shard_id shard, unsigned n, const schema_ptr& s, size_t min_key_size = 1) {
+    return do_make_keys(n, s, min_key_size, shard);
+}
+
+inline sstring make_key_for_shard(shard_id shard, const schema_ptr& s, size_t min_key_size = 1) {
+    return do_make_keys(1, s, min_key_size, shard).front();
+}
+
+inline std::vector<sstring> make_local_keys(unsigned n, const schema_ptr& s, size_t min_key_size = 1) {
+    return do_make_keys(n, s, min_key_size, local_shard_only::yes);
+}
+
+inline sstring make_local_key(const schema_ptr& s, size_t min_key_size = 1) {
+    return do_make_keys(1, s, min_key_size, local_shard_only::yes).front();
+}
+
+inline std::vector<sstring> make_keys(unsigned n, const schema_ptr& s, size_t min_key_size = 1) {
+    return do_make_keys(n, s, min_key_size, local_shard_only::no);
+}
 
 // Helper for working with the following table:
 //
