+++ test/lib/sstable_utils.cc
@@ -13,9 +13,7 @@
 #include "dht/i_partitioner.hh"
 #include "dht/murmur3_partitioner.hh"
 #include <boost/range/irange.hpp>
-#include <boost/range/adaptor/map.hpp>
 #include <boost/range/adaptor/reversed.hpp>
-#include <boost/range/algorithm/sort.hpp>
 #include "sstables/version.hh"
 #include "test/lib/flat_mutation_reader_assertions.hh"
 #include "test/lib/reader_concurrency_semaphore.hh"
@@ -25,35 +23,6 @@
 
 using namespace sstables;
 using namespace std::chrono_literals;
-
-std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size, std::optional<shard_id> shard) {
-    std::vector<std::pair<sstring, dht::decorated_key>> p;
-    p.reserve(n);
-
-    auto key_id = 0U;
-    auto generated = 0U;
-    while (generated < n) {
-        auto raw_key = sstring(std::max(min_key_size, sizeof(key_id)), int8_t(0));
-        std::copy_n(reinterpret_cast<int8_t*>(&key_id), sizeof(key_id), raw_key.begin());
-        auto dk = dht::decorate_key(*s, partition_key::from_single_value(*s, to_bytes(raw_key)));
-        key_id++;
-        if (shard) {
-            if (*shard != shard_of(*s, dk.token())) {
-                continue;
-            }
-        }
-        generated++;
-        p.emplace_back(std::move(raw_key), std::move(dk));
-    }
-    boost::sort(p, [&] (auto& p1, auto& p2) {
-        return p1.second.less_compare(*s, p2.second);
-    });
-    return boost::copy_range<std::vector<sstring>>(p | boost::adaptors::map_keys);
-}
-
-std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size, local_shard_only lso) {
-    return do_make_keys(n, s, min_key_size, lso ? std::optional(this_shard_id()) : std::nullopt);
-}
 
 sstables::shared_sstable make_sstable_containing(std::function<sstables::shared_sstable()> sst_factory, std::vector<mutation> muts) {
     tests::reader_concurrency_semaphore_wrapper semaphore;
