+++ test/lib/test_utils.cc
@@ -8,10 +8,12 @@
 
 #include "test/lib/test_utils.hh"
 
+#include <boost/range/adaptor/map.hpp>
+#include <boost/range/algorithm/sort.hpp>
 #include <seastar/core/print.hh>
 #include <seastar/util/backtrace.hh>
 #include "test/lib/log.hh"
-
+#include "test/lib/simple_schema.hh"
 #include "seastarx.hh"
 #include <random>
 
@@ -69,4 +71,33 @@ sstring make_random_numeric_string(size_t size) {
         b = dist(rng);
     }
     return str;
+}
+
+std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size, std::optional<shard_id> shard) {
+    std::vector<std::pair<sstring, dht::decorated_key>> p;
+    p.reserve(n);
+
+    auto key_id = 0U;
+    auto generated = 0U;
+    while (generated < n) {
+        auto raw_key = sstring(std::max(min_key_size, sizeof(key_id)), int8_t(0));
+        std::copy_n(reinterpret_cast<int8_t*>(&key_id), sizeof(key_id), raw_key.begin());
+        auto dk = dht::decorate_key(*s, partition_key::from_single_value(*s, to_bytes(raw_key)));
+        key_id++;
+        if (shard) {
+            if (*shard != shard_of(*s, dk.token())) {
+                continue;
+            }
+        }
+        generated++;
+        p.emplace_back(std::move(raw_key), std::move(dk));
+    }
+    boost::sort(p, [&] (auto& p1, auto& p2) {
+        return p1.second.less_compare(*s, p2.second);
+    });
+    return boost::copy_range<std::vector<sstring>>(p | boost::adaptors::map_keys);
+}
+
+std::vector<sstring> do_make_keys(unsigned n, const schema_ptr& s, size_t min_key_size, local_shard_only lso) {
+    return do_make_keys(n, s, min_key_size, lso ? std::optional(this_shard_id()) : std::nullopt);
 }
