+++ test/perf/perf_sstable.hh
@@ -14,6 +14,7 @@
 #include "compaction/compaction_manager.hh"
 #include "compaction/time_window_compaction_strategy.hh"
 #include "cell_locking.hh"
+#include "test/lib/simple_schema.hh"
 #include "test/lib/sstable_utils.hh"
 #include "test/lib/test_services.hh"
 #include "test/lib/random_utils.hh"
